backend/requirements.txt
amqp==5.0.6
asgiref==3.3.4
autopep8==1.5.7
billiard==3.6.4.0
celery==5.0.5
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.25.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.3
django-appconf==1.0.4
django-celery-email==3.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
flake8==3.9.2
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sentry-sdk==1.4.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
